+++ packages/fixie/src/agent.ts
@@ -2,10 +2,11 @@ import { gql } from '@apollo/client/core/index.js';
 import yaml from 'js-yaml';
 import fs from 'fs';
 import terminal from 'terminal-kit';
-import { execSync } from 'child_process';
+import { execSync, ChildProcess } from 'child_process';
 import ora from 'ora';
 import os from 'os';
 import path from 'path';
+import { execa } from 'execa';
 
 const { terminal: term } = terminal;
 
@@ -30,7 +31,7 @@ export interface AgentConfig {
   name?: string;
   description?: string;
   moreInfoUrl?: string;
-  public?: boolean;
+  deploymentUrl?: string;
 }
 
 /** Represents metadata about an agent revision. */
@@ -55,7 +56,10 @@ export class FixieAgent {
 
   /** Return the URL for this agent's page on Fixie. */
   public agentUrl(): string {
-    return `${this.client.url}/agents/${this.agentId}`;
+    // TODO(mdw): We need a way to know what the appropriate 'console' URL
+    // is for a given API URL. Since for now this is always console.fixie.ai,
+    // we can just hardcode it.
+    return `https://console.fixie.ai/agents/${this.agentId}`;
   }
 
   /** Get the agent with the given agent ID. */
@@ -77,12 +81,7 @@ export class FixieAgent {
         }
       `,
     });
-    return Promise.all(
-      result.data.allAgentsForUser.map(async (agent: any) => {
-        const retAgent = await this.GetAgent(client, agent.agentId);
-        return retAgent;
-      })
-    );
+    return Promise.all(result.data.allAgentsForUser.map((agent: any) => this.GetAgent(client, agent.agentId)));
   }
 
   /** Return the metadata associated with the given agent. */
@@ -154,7 +153,7 @@ export class FixieAgent {
         name,
         description,
         moreInfoUrl,
-        published: published ?? false,
+        published: published ?? true,
       },
     });
     const agentId = result.data.createAgent.agent.agentId;
@@ -226,14 +225,32 @@ export class FixieAgent {
 
   /** Load an agent configuration from the given directory. */
   public static LoadConfig(agentPath: string): AgentConfig {
-    const config = yaml.load(fs.readFileSync(`${agentPath}/agent.yaml`, 'utf8')) as AgentConfig;
-    return config;
+    const fullPath = path.resolve(path.join(agentPath, 'agent.yaml'));
+    const config = yaml.load(fs.readFileSync(fullPath, 'utf8')) as object;
+
+    // Warn if any fields are present in config that are not supported.
+    const validKeys = [
+      'handle',
+      'name',
+      'description',
+      'moreInfoUrl',
+      'more_info_url',
+      'deploymentUrl',
+      'deployment_url',
+    ];
+    const invalidKeys = Object.keys(config).filter((key) => !validKeys.includes(key));
+    for (const key of invalidKeys) {
+      term('❓ Ignoring invalid key ').yellow(key)(' in agent.yaml\n');
+    }
+    return config as AgentConfig;
   }
 
   /** Package the code in the given directory and return the path to the tarball. */
   private static getCodePackage(agentPath: string): string {
     // Read the package.json file to get the package name and version.
-    const packageJson = JSON.parse(fs.readFileSync(`${agentPath}/package.json`, 'utf8'));
+
+    const packageJsonPath = path.resolve(path.join(agentPath, 'package.json'));
+    const packageJson = JSON.parse(fs.readFileSync(packageJsonPath, 'utf8'));
 
     // Create a temporary directory and run `npm pack` inside.
     const tempdir = fs.mkdtempSync(path.join(os.tmpdir(), `fixie-tmp-${packageJson.name}-${packageJson.version}-`));
@@ -243,24 +260,33 @@ export class FixieAgent {
   }
 
   /** Create a new agent revision, which deploys the agent. */
-  private async createRevision(tarball: string, environmentVariables: Record<string, string>): Promise<string> {
-    const uploadFile = fs.readFileSync(fs.realpathSync(tarball));
+  private async createRevision({
+    externalUrl,
+    tarball,
+    environmentVariables,
+  }: {
+    externalUrl?: string;
+    tarball?: string;
+    environmentVariables: Record<string, string>;
+  }): Promise<string> {
+    const uploadFile = tarball ? fs.readFileSync(fs.realpathSync(tarball)) : undefined;
 
     const result = await this.client.gqlClient().mutate({
       mutation: gql`
         mutation CreateAgentRevision(
           $handle: String!
-          $codePackage: Upload!
-          $environmentVariables: [EnvironmentVariableInput!]
+          $metadata: [RevisionMetadataKeyValuePairInput!]!
+          $makeCurrent: Boolean!
+          $externalDeployment: ExternalDeploymentInput
+          $managedDeployment: ManagedDeploymentInput
         ) {
           createAgentRevision(
             agentHandle: $handle
+            makeCurrent: $makeCurrent
             revision: {
-              managedDeployment: {
-                environment: NODEJS
-                codePackage: $codePackage
-                environmentVariables: $environmentVariables
-              }
+              metadata: $metadata
+              externalDeployment: $externalDeployment
+              managedDeployment: $managedDeployment
             }
           ) {
             revision {
@@ -271,57 +297,20 @@ export class FixieAgent {
       `,
       variables: {
         handle: this.handle,
-        codePackage: new Blob([uploadFile], { type: 'application/gzip' }),
-        environmentVariables: Object.entries(environmentVariables).map(([key, value]) => ({ name: key, value })),
+        metadata: [],
+        makeCurrent: true,
+        externalDeployment: externalUrl && { url: externalUrl },
+        managedDeployment: tarball &&
+          uploadFile && {
+            environment: 'NODEJS',
+            codePackage: new Blob([uploadFile], { type: 'application/gzip' }),
+            environmentVariables: Object.entries(environmentVariables).map(([key, value]) => ({ name: key, value })),
+          },
       },
       fetchPolicy: 'no-cache',
     });
 
     return result.data.createAgentRevision.revision.id;
-  }
-
-  /** Deploy an agent from the given directory. */
-  public static async DeployAgent(
-    client: FixieClient,
-    agentPath: string,
-    environmentVariables: Record<string, string> = {}
-  ): Promise<string> {
-    const config = await FixieAgent.LoadConfig(agentPath);
-    const agentId = `${(await client.userInfo()).username}/${config.handle}`;
-    term('🦊 Deploying agent ').green(agentId)('...\n');
-
-    // Check if the package.json path exists in this directory.
-    if (!fs.existsSync(`${agentPath}/package.json`)) {
-      throw Error(`No package.json found in ${agentPath}. Only JS-based agents are supported.`);
-    }
-
-    let agent: FixieAgent;
-    try {
-      agent = await FixieAgent.GetAgent(client, agentId);
-      term('👽 Updating agent ').green(agentId)('...\n');
-      agent.update({
-        name: config.name,
-        description: config.description,
-        moreInfoUrl: config.moreInfoUrl,
-        published: config.public,
-      });
-    } catch (e) {
-      // Try to create the agent instead.
-      term('🌲 Creating new agent ').green(agentId)('...\n');
-      agent = await FixieAgent.CreateAgent(
-        client,
-        config.handle,
-        config.name,
-        config.description,
-        config.moreInfoUrl,
-        config.public
-      );
-    }
-    const tarball = FixieAgent.getCodePackage(agentPath);
-    const spinner = ora(' 🚀 Deploying...').start();
-    const revision = await agent.createRevision(tarball, environmentVariables);
-    spinner.succeed(`Revision ${revision} was deployed to ${agent.agentUrl()}`);
-    return revision;
   }
 
   /** Get the current agent revision. */
@@ -342,5 +331,213 @@ export class FixieAgent {
     });
 
     return result.data.agentById.currentRevision as AgentRevision;
+  }
+
+  /** Ensure that the agent is created or updated. */
+  private static async ensureAgent(client: FixieClient, agentId: string, config: AgentConfig): Promise<FixieAgent> {
+    let agent: FixieAgent;
+    try {
+      agent = await FixieAgent.GetAgent(client, agentId);
+      agent.update({
+        name: config.name,
+        description: config.description,
+        moreInfoUrl: config.moreInfoUrl,
+      });
+    } catch (e) {
+      // Try to create the agent instead.
+      term('🦊 Creating new agent ').green(agentId)('...\n');
+      agent = await FixieAgent.CreateAgent(client, config.handle, config.name, config.description, config.moreInfoUrl);
+    }
+    return agent;
+  }
+
+  static spawnAgentProcess(agentPath: string, port: number): ChildProcess {
+    term(`🌱 Starting local agent process on port ${port}...\n`);
+    const pathToCheck = path.resolve(path.join(agentPath, 'dist', 'index.js'));
+    if (!fs.existsSync(pathToCheck)) {
+      throw Error(`Your agent was not found at ${pathToCheck}. Do you need to build your agent code first?`);
+    }
+
+    const cmdline = `npx --package=@fixieai/sdk fixie-serve-bin --packagePath ./dist/index.js --port ${port}`;
+    // Split cmdline into the first value (argv0) and a list of arguments separated by spaces.
+    term('🌱 Running: ').green(cmdline)('\n');
+
+    const [argv0, ...args] = cmdline.split(' ');
+    const subProcess = execa(argv0, args, { cwd: agentPath });
+
+    subProcess.stdout?.on('data', (sdata: string) => {
+      console.log(`🌱 Agent stdout: ${sdata}`);
+    });
+    subProcess.stderr?.on('data', (sdata: string) => {
+      console.error(`🌱 Agent stdout: ${sdata}`);
+    });
+    subProcess.on('close', (returnCode: number) => {
+      term(`🌱 Agent child process exited with code ${returnCode} - `).red('exiting.\n');
+      process.exit(returnCode);
+    });
+    return subProcess;
+  }
+
+  /** Deploy an agent from the given directory. */
+  public static async DeployAgent(
+    client: FixieClient,
+    agentPath: string,
+    environmentVariables: Record<string, string> = {}
+  ): Promise<string> {
+    const config = await FixieAgent.LoadConfig(agentPath);
+    const agentId = `${(await client.userInfo()).username}/${config.handle}`;
+    term('🦊 Deploying agent ').green(agentId)('...\n');
+
+    // Check that the package.json path exists in this directory.
+    const packageJsonPath = path.resolve(path.join(agentPath, 'package.json'));
+    if (!fs.existsSync(packageJsonPath)) {
+      throw Error(`No package.json found at ${packageJsonPath}. Only JS-based agents are supported.`);
+    }
+    const agent = await this.ensureAgent(client, agentId, config);
+    const tarball = FixieAgent.getCodePackage(agentPath);
+    const spinner = ora(' 🚀 Deploying... (hang tight, this takes a minute or two!)').start();
+    const revision = await agent.createRevision({ tarball, environmentVariables });
+    spinner.succeed(`Agent ${config.handle} is running at: ${agent.agentUrl()}`);
+    return revision;
+  }
+
+  /** Run an agent locally from the given directory. */
+  public static async ServeAgent({
+    client,
+    agentPath,
+    tunnel,
+    port,
+    environmentVariables,
+    debug,
+  }: {
+    client: FixieClient;
+    agentPath: string;
+    tunnel?: boolean;
+    port: number;
+    environmentVariables: Record<string, string>;
+    debug?: boolean;
+  }) {
+    const config = await FixieAgent.LoadConfig(agentPath);
+    const agentId = `${(await client.userInfo()).username}/${config.handle}`;
+    term('🦊 Serving agent ').green(agentId)('...\n');
+
+    // Check if the package.json path exists in this directory.
+    const packageJsonPath = path.resolve(path.join(agentPath, 'package.json'));
+    if (!fs.existsSync(packageJsonPath)) {
+      throw Error(`No package.json found in ${packageJsonPath}. Only JS-based agents are supported.`);
+    }
+
+    // Start the agent process locally.
+    const agentProcess = FixieAgent.spawnAgentProcess(agentPath, port);
+
+    // Wait for 5 seconds for it to start up.
+    await new Promise((resolve) => setTimeout(resolve, 5000));
+
+    async function* spawnTunnel(port: number): AsyncGenerator<string> {
+      type AsyncGeneratorYield<T> = (value: T) => void;
+      // Represents the value yielded by the subprocess.
+      let yieldValue: AsyncGeneratorYield<string> | null = null;
+      // This Promise is resolved when the tunnel subprocess yields a new address.
+      let whenYielded = new Promise<string>((resolve) => {
+        yieldValue = resolve;
+      });
+
+      term('🚇 Starting tunnel process...\n');
+      // We use localhost.run as a tunneling service. This sets up an SSH tunnel
+      // to the provided local port via localhost.run. The subprocess returns a
+      // stream of JSON responses, one per line, with the external URL of the tunnel
+      // as it changes.
+      const subProcess = execa('ssh', [
+        '-R',
+        // N.B. 127.0.0.1 must be used on Windows (not localhost or 0.0.0.0)
+        `80:127.0.0.1:${port}`,
+        '-o',
+        // Need to send keepalives to prevent the connection from getting chopped
+        // (see https://localhost.run/docs/faq#my-connection-is-unstable-tunnels-go-down-often)
+        'ServerAliveInterval=59',
+        '-o',
+        'StrictHostKeyChecking=accept-new',
+        'nokey@localhost.run',
+        '--',
+        '--output=json',
+      ]);
+      subProcess.stdout?.setEncoding('utf8');
+
+      // Every time the subprocess emits a new line, we parse it as JSON ans
+      // extract the 'address' field.
+      let currentLine = '';
+      subProcess.stdout?.on('data', (chunk: string) => {
+        // We need to do buffering since the data we get from stdout
+        // will not necessarily be line-buffered. We can get 0, 1, or more complete
+        // lines in a single chunk.
+        currentLine += chunk;
+        let newlineIndex;
+        while ((newlineIndex = currentLine.indexOf('\n')) !== -1) {
+          const line = currentLine.slice(0, newlineIndex);
+          currentLine = currentLine.slice(newlineIndex + 1);
+          // Parse data as JSON.
+          const pdata = JSON.parse(line);
+          // If pdata has the 'address' field, yield it.
+          if (pdata.address) {
+            yieldValue?.(`https://${pdata.address}`);
+          }
+        }
+      });
+      if (debug) {
+        subProcess.stderr?.on('data', (sdata: string) => {
+          console.error(`🚇 Tunnel stderr: ${sdata}`);
+        });
+        subProcess.on('close', (returnCode: number) => {
+          console.log(`🚇 Tunnel child process exited with code ${returnCode}`);
+        });
+      }
+      // Now we wait for the subprocess to yield a new address, which we then
+      // re-yield to the caller of this function.
+      while (true) {
+        const tunnelAddress = await whenYielded;
+        if (tunnelAddress === '') {
+          break;
+        }
+        yield tunnelAddress;
+        yieldValue = null;
+        whenYielded = new Promise<string>((resolve) => {
+          yieldValue = resolve;
+        });
+      }
+    }
+
+    // This is an iterator which yields the public URL of the tunnel where the agent
+    // can be reached by the Fixie service. The tunnel address can change over time.
+    let deploymentUrlsIter;
+    if (tunnel) {
+      deploymentUrlsIter = spawnTunnel(port);
+    } else {
+      if (!config.deploymentUrl) {
+        throw Error('No deployment URL specified in agent.yaml');
+      }
+      deploymentUrlsIter = [
+        (async function* () {
+          yield config.deploymentUrl;
+        })(),
+      ];
+    }
+
+    // TODO(mdw): We need to restore the original agent deployment (if one existed)
+    // after the process finishes.
+    const agent = await this.ensureAgent(client, agentId, config);
+
+    for await (const currentUrl of deploymentUrlsIter) {
+      term('🚇 Current tunnel URL is: ').green(currentUrl)('\n');
+      try {
+        // Wait 3 seconds to ensure the tunnel is set up.
+        await new Promise((resolve) => setTimeout(resolve, 3000));
+        await agent.createRevision({ externalUrl: currentUrl as string, environmentVariables });
+        term('🥡 Agent ').green(config.handle)(' is running at: ').green(agent.agentUrl())('\n');
+      } catch (e: any) {
+        term('🥡 Got error trying to create agent revision: ').red(e.message)('\n');
+        console.error(e);
+        continue;
+      }
+    }
   }
 }