+++ packages/fixie/src/client.ts
@@ -1,81 +1,29 @@
-import type { Jsonifiable } from 'type-fest';
 import { ApolloClient } from '@apollo/client/core/ApolloClient.js';
 import { InMemoryCache } from '@apollo/client/cache/inmemory/inMemoryCache.js';
 import createUploadLink from 'apollo-upload-client/public/createUploadLink.js';
 import isExtractableFile from 'apollo-upload-client/public/isExtractableFile.js';
 import { ExtractableFile } from 'extract-files';
 import { ReadStream } from 'fs';
-
-export interface UserInfo {
-  id: number;
-  username: string;
-  is_authenticated: boolean;
-  is_superuser: boolean;
-  is_staff: boolean;
-  is_active: boolean;
-  is_anonymous: boolean;
-  email?: string;
-  first_name?: string;
-  last_name?: string;
-  last_login?: Date;
-  date_joined?: Date;
-  api_token?: string;
-  avatar?: string;
-  organization?: string;
-}
+import { IsomorphicFixieClient } from './isomorphic-client.js';
 
 /**
  * A client to the Fixie AI platform.
+ *
+ * This client can only be used in NodeJS.
  */
-export class FixieClient {
-  url!: string;
-  apiKey!: string;
-
-  /** Use Create() instead. */
-  private constructor() {}
-
-  /**
-   * Create a new Fixie client.
-   * @param url The URL of the Fixie API server.
-   */
-  public static Create(url: string): FixieClient {
-    const client = new FixieClient();
-    client.url = url;
-    const apiKey = process.env.FIXIE_API_KEY;
-    if (!apiKey) {
+export class FixieClient extends IsomorphicFixieClient {
+  static Create(url: string, apiKey?: string): FixieClient {
+    const apiKeyToUse = apiKey ?? process.env.FIXIE_API_KEY;
+    if (!apiKeyToUse) {
       throw new Error(
-        'You must set the FIXIE_API_KEY environment variable. This can be found at: https://beta.fixie.ai/profile'
+        'You must pass apiKey to the constructor, or set the FIXIE_API_KEY environment variable. The API key can be found at: https://console.fixie.ai/profile'
       );
     }
-    client.apiKey = apiKey;
-    return client;
+    return new this(url, apiKeyToUse);
   }
 
-  /** Send a request to the Fixie API with the appropriate auth headers. */
-  async request(path: string, bodyData?: any): Promise<Jsonifiable> {
-    let res;
-    if (bodyData) {
-      const body = JSON.stringify(bodyData);
-      res = await fetch(`${this.url}${path}`, {
-        method: 'POST',
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-          'Content-Type': 'application/json',
-        },
-        body,
-      });
-    } else {
-      res = await fetch(`${this.url}${path}`, {
-        method: 'GET',
-        headers: {
-          Authorization: `Bearer ${this.apiKey}`,
-        },
-      });
-    }
-    if (!res.ok) {
-      throw new Error(`Failed to access Fixie API ${this.url}${path}: ${res.statusText}`);
-    }
-    return res.json();
+  static CreateWithoutApiKey(url: string): FixieClient {
+    return new this(url);
   }
 
   /** Return a GraphQL client for the Fixie API. */
@@ -96,122 +44,5 @@ export class FixieClient {
           isExtractableFile(value) || (typeof ReadStream !== 'undefined' && value instanceof ReadStream),
       }),
     });
-  }
-
-  /** Return information on the currently logged-in user. */
-  userInfo(): Promise<UserInfo> {
-    const rawUserInfo: unknown = this.request('/api/user');
-    return rawUserInfo as Promise<UserInfo>;
-  }
-
-  /** List Corpora visible to this user. */
-  listCorpora(): Promise<Jsonifiable> {
-    return this.request('/api/v1/corpora');
-  }
-
-  /** Get information about a given Corpus. */
-  getCorpus(corpusId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}`);
-  }
-
-  /** Create a new Corpus. */
-  createCorpus(name?: string): Promise<Jsonifiable> {
-    const body = {
-      corpus: {
-        display_name: name,
-      },
-    };
-    return this.request('/api/v1/corpora', body);
-  }
-
-  /** Query a given Corpus. */
-  queryCorpus(corpusId: string, query: string, maxChunks?: number): Promise<Jsonifiable> {
-    const body = {
-      corpus_id: corpusId,
-      query,
-      max_chunks: maxChunks,
-    };
-    return this.request(`/api/v1/corpora/${corpusId}:query`, body);
-  }
-
-  /** List the Sources in a given Corpus. */
-  listCorpusSources(corpusId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}/sources`);
-  }
-
-  /** Get information about a given Source. */
-  getCorpusSource(corpusId: string, sourceId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}/sources/${sourceId}`);
-  }
-
-  /** Add a new Source to a Corpus. */
-  addCorpusSource(corpusId: string, urlPattern: string): Promise<Jsonifiable> {
-    const body = {
-      corpus_id: corpusId,
-      source: {
-        corpus_id: corpusId,
-        load_spec: { web: { start_urls: [urlPattern] } },
-        process_steps: [
-          {
-            step_name: 'markdownify',
-            relevant_document_types: {
-              include: { mime_types: ['text/html'] },
-            },
-            html_to_markdown: {},
-          },
-        ],
-        embed_specs: [
-          {
-            spec_name: 'markdown',
-            relevant_document_types: {
-              include: { mime_types: ['text/markdown'] },
-            },
-            max_chunk_size: 1000,
-            chunk_overlap: 100,
-            splits: ['\n\n', '\n', ' ', ''],
-          },
-          {
-            spec_name: 'plain',
-            relevant_document_types: {
-              include: { mime_types: ['text/plain'] },
-            },
-            max_chunk_size: 1000,
-            chunk_overlap: 100,
-            splits: ['\n\n', '\n', ' ', ''],
-          },
-        ],
-      },
-    };
-
-    return this.request(`/api/v1/corpora/${corpusId}/sources`, body);
-  }
-
-  /** Refresh the given Source. */
-  refreshCorpusSource(corpusId: string, sourceId: string): Promise<Jsonifiable> {
-    const body = {
-      corpus_id: corpusId,
-      source_id: sourceId,
-    };
-    return this.request(`/api/v1/corpora/${corpusId}/sources/${sourceId}:refresh`, body);
-  }
-
-  /** List Jobs associated with a given Source. */
-  listCorpusSourceJobs(corpusId: string, sourceId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}/sources/${sourceId}/jobs`);
-  }
-
-  /** Get information about a given Job. */
-  getCorpusSourceJob(corpusId: string, sourceId: string, jobId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}/sources/${sourceId}/jobs/${jobId}`);
-  }
-
-  /** List Documents in a given Corpus. */
-  listCorpusDocs(corpusId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}/documents`);
-  }
-
-  /** Get information about a given Document. */
-  getCorpusDoc(corpusId: string, docId: string): Promise<Jsonifiable> {
-    return this.request(`/api/v1/corpora/${corpusId}/documents/${docId}`);
   }
 }